fix: 모티티 삭제 메서드 중복 제거

+++ src/main/java/com/motivation/mojaty/domain/motivation/service/MotivationService.java
@@ -96,7 +96,9 @@ public class MotivationService {
 
         validateMotivationFromUser(motivation);
 
-        fileService.deleteFile(motivation.getContent());
+        if(motivation.getContent().contains(FileService.IMAGE_URL)) {
+            fileService.deleteFile(motivation.getContent());
+        }
         motivationRepository.delete(motivation);
     }
